refactor: share headless authentication prompts
